git subrepo pull projects/octopus
subrepo:
  subdir:   "projects/octopus"
  merged:   "8ad0c17"
upstream:
  origin:   "git@github.com:octopus-platform/octopus.git"
  branch:   "master"
  commit:   "8ad0c17"
git-subrepo:
  version:  "0.3.0"
  origin:   "https://github.com/ingydotnet/git-subrepo.git"
  commit:   "1e79595"

+++ projects/octopus/octopus-server/src/main/java/octopus/server/components/projectmanager/ProjectManager.java
@@ -1,12 +1,12 @@
 package octopus.server.components.projectmanager;
 
 import com.orientechnologies.orient.client.remote.OServerAdmin;
-import org.apache.commons.io.FileUtils;
 import orientdbimporter.Constants;
 
-import java.io.File;
 import java.io.IOException;
+import java.nio.file.DirectoryStream;
 import java.nio.file.Files;
+import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.Map;
@@ -14,39 +14,45 @@ import java.util.Map;
 public class ProjectManager
 {
 
-	private static String projectsDir;
+	private static Path projectsDir;
 	private static Map<String, OctopusProject> nameToProject = new HashMap<String, OctopusProject>();
 
-	public static void setProjectDir(String newProjectsDir) throws IOException
+	public static void setProjectDir(Path newProjectsDir) throws IOException
 	{
-		projectsDir = new File(newProjectsDir).getCanonicalPath();
+		if (!newProjectsDir.isAbsolute())
+		{
+			newProjectsDir = newProjectsDir.toAbsolutePath();
+		}
+		projectsDir = newProjectsDir.normalize();
 		openProjectsDir();
 		loadProjects();
 	}
 
-	private static void openProjectsDir()
+	private static void openProjectsDir() throws IOException
 	{
-		if (Files.notExists(Paths.get(projectsDir)))
+		if (Files.notExists(projectsDir))
 		{
-			new File(projectsDir).mkdirs();
+			Files.createDirectories(projectsDir);
 		}
 	}
 
-	private static void loadProjects()
+	private static void loadProjects() throws IOException
 	{
-		File projectsDirHandle = new File(projectsDir);
-		File[] files = projectsDirHandle.listFiles();
-		for (File projectDir : files)
+		try (DirectoryStream<Path> stream = Files.newDirectoryStream(projectsDir))
 		{
-			if (!projectDir.isDirectory())
-				continue;
-			loadProject(projectDir);
+			for (Path path : stream)
+			{
+				if (Files.isDirectory(path))
+				{
+					loadProject(path);
+				}
+			}
 		}
 	}
 
-	private static void loadProject(File projectDir)
+	private static void loadProject(Path projectDir) throws IOException
 	{
-		String projectName = projectDir.getName();
+		String projectName = projectDir.getFileName().toString();
 		OctopusProject newProject = createOctopusProjectForName(projectName);
 		nameToProject.put(projectName, newProject);
 	}
@@ -61,15 +67,18 @@ public class ProjectManager
 		return nameToProject.get(name);
 	}
 
-	public static String getPathToProject(String name)
-	{
-		return projectsDir + File.separator + name;
-	}
-
-	public static void create(String name)
+	public static Path getPathToProject(String name)
 	{
 		if (projectsDir == null)
-			throw new RuntimeException("Error: projectDir not set");
+			throw new IllegalStateException("Error: projectDir not set");
+
+		return Paths.get(projectsDir.toString(), name);
+	}
+
+	public static void create(String name) throws IOException
+	{
+		if (projectsDir == null)
+			throw new IllegalStateException("Error: projectDir not set");
 
 		if (doesProjectExist(name))
 			throw new RuntimeException("Project already exists");
@@ -78,10 +87,10 @@ public class ProjectManager
 		nameToProject.put(name, project);
 	}
 
-	public static void delete(String name)
+	public static void delete(String name) throws IOException
 	{
 		if (projectsDir == null)
-			throw new RuntimeException("Error: projectDir not set");
+			throw new IllegalStateException("Error: projectDir not set");
 
 		deleteProjectWithName(name);
 	}
@@ -89,41 +98,34 @@ public class ProjectManager
 	public static Iterable<String> listProjects()
 	{
 		return nameToProject.keySet();
-
 	}
 
-	private static OctopusProject createOctopusProjectForName(String name)
+	private static OctopusProject createOctopusProjectForName(String name) throws IOException
 	{
-		String pathToProject = getPathToProject(name);
-		File dir = new File(pathToProject);
-		dir.mkdirs();
+		Path pathToProject = getPathToProject(name);
+		Files.createDirectories(pathToProject);
 
-		OctopusProject newProject = new OctopusProject();
-		newProject.setPathToProjectDir(pathToProject);
-		newProject.setDatabaseName(name);
+		OctopusProject newProject = new OctopusProject(pathToProject.toString(), name);
 		return newProject;
 	}
 
-	private static void deleteProjectWithName(String name)
+	private static void deleteProjectWithName(String name) throws IOException
 	{
-		File dir = new File(getPathToProject(name));
-		try
-		{
-			FileUtils.deleteDirectory(dir);
-			nameToProject.remove(name);
-			removeDatabase(name);
-		} catch (IOException e)
-		{
-			throw new RuntimeException("IO Exception on delete");
-		}
+		Path pathToProject = getPathToProject(name);
+		Files.deleteIfExists(pathToProject);
+		nameToProject.remove(name);
+		removeDatabaseIfExists(name);
 	}
 
-	private static void removeDatabase(String dbName) throws IOException
+	private static void removeDatabaseIfExists(String dbName) throws IOException
 	{
 		OServerAdmin admin;
 		admin = new OServerAdmin("localhost/" + dbName).connect(
 				Constants.DB_USERNAME, Constants.DB_PASSWORD);
-		admin.dropDatabase("plocal");
+		if (admin.existsDatabase())
+		{
+			admin.dropDatabase("plocal");
+		}
 	}
 
 }
